#309: move PendingFilesRepository to domain transfer package
