+++ composeApp/src/commonMain/kotlin/com/tubetoast/tether/presentation/banners/PendingOutboundBanner.kt
@@ -4,8 +4,8 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewParameter
-import com.tubetoast.tether.presentation.PendingFilesSummary
 import com.tubetoast.tether.transfer.ByteFormatting
+import com.tubetoast.tether.transfer.PendingFilesSummary
 import com.tubetoast.tether.ui.designsystem.Banner
 import com.tubetoast.tether.ui.designsystem.BannerSeverity
 import com.tubetoast.tether.ui.designsystem.Button
