+++ composeApp/src/commonMain/kotlin/com/tubetoast/tether/transfer/PendingFilesRepository.kt
@@ -1,7 +1,5 @@
-package com.tubetoast.tether.presentation.transfer
+package com.tubetoast.tether.transfer
 
-import com.tubetoast.tether.presentation.PendingFilesSummary
-import com.tubetoast.tether.transfer.FileSource
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.update
