+++ composeApp/src/commonMain/kotlin/com/tubetoast/tether/transfer/PendingFilesSummary.kt
@@ -1,4 +1,4 @@
-package com.tubetoast.tether.presentation
+package com.tubetoast.tether.transfer
 
 data class PendingFilesSummary(
     val fileCount: Int,
